requirements.txt
numpy
matplotlib
parselmouth
librosa
pytorch
h5py
tqdm
dareblopy